+++ source/filterMap.js
@@ -1,0 +1,5 @@
+import {mapFn} from './map.js'
+
+export function filterMap(fn) {
+  return list => mapFn(fn, list).filter(Boolean)
+}