+++ source/middle.js
@@ -1,0 +1,6 @@
+import { init } from './init.js'
+import { tail } from './tail.js'
+
+export function middle(listOrString) {
+  return tail(init(listOrString))
+}